move rest help functions into utils module
